fix(training): allow defined precision to map to sequence tensor

+++ src/dnadiffusion/utils/train_util.py
@@ -38,7 +38,7 @@ def train_step(
         dtype = torch.float32
 
     with torch.autocast(device_type=device, dtype=dtype):
-        x = x.to(device)
+        x = x.to(device, dtype=dtype)
         y = y.to(device)
         loss = model(x, y)
